♻️:홈페이지컴포넌트분리
